- refactor some more DB definitions into a separate folder - create a processed field group to mirror the structure of field groups
