+++ src/database/processed_fields/processed_field.py
@@ -12,13 +12,13 @@ class ProcessedField(MappedAsDataclass, OrmBase):
     __tablename__ = "processed_field"
 
     id: Mapped[int] = mapped_column(init=False, primary_key=True)
-    processed_region_id: Mapped[int] = mapped_column(ForeignKey("processed_region.id"), init=False)
 
     processing_error: Mapped[bool]
 
     # Relationships
 
-    processed_region: Mapped["ProcessedRegion"] = relationship(init=False, back_populates="fields")
+    processed_group_id: Mapped[int] = mapped_column(ForeignKey("processed_field_group.id"), init=False)
+    processed_group: Mapped["ProcessedFieldGroup"] = relationship(init=False, back_populates="fields")
 
     text_field: Mapped[ProcessedTextField] = relationship(default=None, back_populates="processed_field")
     checkbox_field: Mapped[ProcessedCheckboxField] = relationship(default=None, back_populates="processed_field")
