+++ src/database/processing/processed_region.py
@@ -1,15 +1,14 @@
 from sqlalchemy import ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, MappedAsDataclass, relationship
 
-from . import OrmBase
-from .processed_fields.processed_field import ProcessedField
+from .. import OrmBase
+from ..processed_fields.processed_field_group import ProcessedFieldGroup
 
 
 class ProcessedRegion(MappedAsDataclass, OrmBase):
     __tablename__ = "processed_region"
 
     id: Mapped[int] = mapped_column(init=False, primary_key=True)
-    process_result_id: Mapped[int] = mapped_column(ForeignKey("process_result.id"), init=False)
 
     local_id: Mapped[int]
     name: Mapped[str]
@@ -19,5 +18,7 @@ class ProcessedRegion(MappedAsDataclass, OrmBase):
 
     # Relationships
 
+    process_result_id: Mapped[int] = mapped_column(ForeignKey("process_result.id"), init=False)
     process_result: Mapped["ProcessResult"] = relationship(init=False, back_populates="regions")
-    fields: Mapped[list[ProcessedField]] = relationship(init=False, back_populates="processed_region")
+
+    groups: Mapped[list[ProcessedFieldGroup]] = relationship(init=False, back_populates="processed_region")
